refactor: unify user creation types
