+++ src/lib/repositories/index.ts
@@ -1,3 +1,3 @@
 export { BaseRepository } from "./base.repository";
-export type { CreateUserInput, UserPublicData } from "./user.repository";
 export { UserRepository } from "./user.repository";
+export type { UserCreate, UserPublicData } from "./user.types";
