+++ src/lib/repositories/user.types.ts
@@ -1,0 +1,16 @@
+export interface UserCreate {
+  name: string;
+  email: string;
+  password: string;
+  defaultHourlyRate?: number;
+  currency?: string;
+}
+
+export interface UserPublicData {
+  id: string;
+  name: string;
+  email: string;
+  defaultHourlyRate: number;
+  currency: string;
+  createdAt: Date;
+}
