+++ src/lib/services/index.ts
@@ -1,2 +1,2 @@
-export type { AuthenticateUserData, CreateUserData, IUserRepository } from "./auth.service";
+export type { AuthenticateUserData, IUserRepository } from "./auth.service";
 export { AuthService } from "./auth.service";
